// generated/addon/angular/src/app/api-models/cieb-country-subdivision.model.ts
/* Copyright (c) 2021 . All Rights Reserved. */

export class CiebCountrySubdivision {
  countryCode2: string;
  countrySubDivCode: string;
  regionalDivision: string;
  subDivCategory: string;
  subDivisionName: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}